requirements.txt
Flask
Flask-WTF
FLask-SQLAlchemy
Flask-Login
gunicorn